subject: single commit my branch has
description: i have a long long description here, that I care a lot about, for sure

description: i have a long long description here, that I care a lot about, for sure

## Summary of Changes

This pull request involves a significant change to the `.gitignore` file.

### Changes Made

The line in the `.gitignore` file that was responsible for ignoring the `.vs` directory has been modified. The previous line was `"/.vs"`, but it has been changed to `"/.vs+"`.

### Reason for Changes

The reason for this change is to make the `.gitignore` file ignore a directory named `.vs+` instead of `.vs`.

### Significant Design Decisions

This is a significant design decision as it alters the directories that are ignored by Git. It's important to note that this change will affect the way the repository interacts with the `.vs+` directory.

Please review the changes and provide your feedback.